MGS-5581 [#imx-1868] update header files of nnrt and OVXLIB
update header files of nnrt and OVXLIB for patch_base_on_6.4.0.p2_from_20200306_20200320

+++ include/OVXLIB/utils/vsi_nn_dtype_util.h
@@ -126,7 +126,7 @@ OVXLIB_API int32_t vsi_nn_Fp32ToAffine
     (
     const float  in,
     const float  scale,
-    const uint8_t    zero_point,
+    const int32_t    zero_point,
     const vsi_nn_type_e type
     );
 
@@ -134,7 +134,7 @@ OVXLIB_API float vsi_nn_AffineToFp32
     (
     const int32_t    val,
     const float  scale,
-    const uint8_t    zero_point,
+    const int32_t    zero_point,
     const vsi_nn_type_e type
     );
